pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signalrgb-magichome-bridge"
version = "1.0.0"
description = "Use SignalRGB to control Magic Home WiFi SPI LED controllers in real-time"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Almog Cohen" }]
keywords = ["signalrgb", "magic-home", "wled", "led", "addressable", "rgb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Home Automation",
]
dependencies = [
    "aiohttp>=3.9,<4",
    "zeroconf>=0.131",
    "flux_led>=1.0",
]

[project.optional-dependencies]
tray = ["pystray>=0.19", "Pillow>=10.0"]
dev = ["pytest", "pytest-aiohttp"]

[project.urls]
Homepage = "https://github.com/RisingStar1/signalrgb-magichome-bridge"
Repository = "https://github.com/RisingStar1/signalrgb-magichome-bridge"
Issues = "https://github.com/RisingStar1/signalrgb-magichome-bridge/issues"

[project.scripts]
signalrgb-bridge = "signalrgb_magichome_bridge.bridge:main"
signalrgb-bridge-tray = "signalrgb_magichome_bridge.tray:main"
